requirements.txt
fastapi==0.109.1
uvicorn==0.24.0
pydantic==2.4.2
openai==1.3.9
python-dotenv==1.0.0
httpx>=0.23.0,<0.25.0
aiohttp>=3.9.0
aiofiles==23.2.1
python-multipart==0.0.19
tenacity==8.2.3
Pillow==10.1.0
uuid==1.30
python-jose==3.4.0